refactor(visitor): simplify repeated logics
