+++ packages/swc-plugin-istanbul-oxi/src/utils/mod.rs
@@ -2,3 +2,4 @@
 
 pub(crate) mod ast_builder;
 pub(crate) mod lookup_range;
+pub(crate) mod visitor_macros;
